refactor: msw 값 반환시에 내림차순으로 정렬
