+++ frontend/src/mocks/storage/Reminder.ts
@@ -3,14 +3,13 @@ import { getParticularDateFromToday } from "utils/date";
 
 const KEY = 'MSW_REMINDER';
 
-const getAll = (): Reminder[] => {
+const getAll = ():{data:Reminder[]} => {
   const storageData = sessionStorage.getItem(KEY);
-  return storageData ? JSON.parse(storageData) : [];
+  return storageData ? JSON.parse(storageData) : {data:[]};
 };
 
 const water = (id:number) =>{
   const {data} = JSON.parse(sessionStorage.getItem(KEY) ?? '[]') as {data:Reminder[]}
-
   const updatedData = data.map((data) => {
     const {petPlantId} = data
     if(id !== petPlantId) return data
@@ -22,7 +21,8 @@ const water = (id:number) =>{
     }
   })
 
-  sessionStorage.setItem(KEY, JSON.stringify(updatedData))
+
+  sessionStorage.setItem(KEY, JSON.stringify({data:updatedData}))
 }
